requirements.txt
discord.py==1.6.0
python-dotenv==0.15.0
requests==2.24.0
pandas==1.1.3
sqlalchemy==1.3.20
psycopg2==2.8.6
pytz==2020.1
pytest==6.2.2
aiocron==1.8